# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample Unicef server."
  version: "1.0.0"
  title: "Unicef"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: ""
basePath: "/"
...
schemes:
- "http"
- "https"
paths:
...
  /api/view/allsuppliers:
    get:
      tags:
      - "Suppliers"
      summary: "All suppliers list"
      description: "This api will give all suppliers list"
      operationId: "fetchSuppliers"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "boolean"
                example: true
              message:
                type: "string"
                example: "Supplier found!!"
              data:
                type: "array"
                items:
                  $ref: "#/definitions/Supplier"
        "500":
          description: "Invalid input"
          schema:
            type: "object"
            properties:
              status:
                type: "boolean"
                example: false
              message:
                type: "string"
                example: "Supplier not found!!"
              data:
                type: "array"
                items:
                  type: string
                  example: 
         
               
      # security:
      # - petstore_auth:
      #   - "write:pets"
      #   - "read:pets"
   
# securityDefinitions:
#   petstore_auth:
#     type: "oauth2"
#     authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
#     flow: "implicit"
#     scopes:
#       write:pets: "modify pets in your account"
#       read:pets: "read your pets"
#   api_key:
#     type: "apiKey"
#     name: "api_key"
#     in: "header"
  /api/add/supplier:
    post:
      tags:
      - "Suppliers"
      summary: "Creates list of supplier with given input"
      description: ""
      operationId: "addSupplier"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "body"
        description: "Supplier object that needs to be added to the database"
        required: true
        schema:
          type: "object"
          properties:
            _partition: 
               type: "string"
               example: "101"
            userId:
              type: "string"
              #format: "objectId"
              example: "6048708c946251d02c88f329"
            addressLine1:
              type: "string"
              example: "PUNE"
            addressLine2:
              type: "string"
              example: "BANNER"
            city:
              type: "string"
              example: "PUNE"
            country:
              type: "string"
              example: "INDIA"
            isoCountry:
              type: "string"
              example: "phl"
            district:
              type: "string"
              example: "PUNE"
            postalCode:
              type: "string"
              example: "784512"
            directions:
              type: "string"
              example: "WEST"
            landmark:
              type: "string"
              example: "NEW GALAXY MALL"
            region:
              type: "string"
              example: "EAST"
            town:
              type: "string"
              example: "PUNE"
            zip:
              type: "string"
              example: "PUNE"
            email:
              type: "string"
              example: "email@gmail.com"
            phone:
              type: "string"
              example: "9999999999"
            supplierCode:
              type: "string"
              required: true
              example: "ALKEM"
            deliveryFee:
              type: "integer"
              format: "decimal"
              example: 100.00
            isoCountry:
              type: "string"
              example: "phl"
            lastProductSeq:
              type: "string"
              example: "45A"
            supplierName:
              type: "string"
              example: "ALKEM"
            catalogTags:
              type: "array"
              xml:
                name: "tag"
                wrapped: true
              items:
                type: "string"
            usdPrice:
              type: "integer"
              format: "decimal"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "boolean"
                example: true
              message:
                type: "string"
                example: "Supplier added successfully"
              supplierId:
                type: "string"
                example: "6048708c946251d02c88f329"
        "500":
          description: "Invalid input"
          schema:
            type: "object"
            properties:
              status:
                type: "boolean"
                example: false
              message:
                type: "string"
                example: "Supplier code already exists!!"
              supplierId:
                type: "string"
                example: "null"
        "400":
          description: "parameter is missing"
          schema:
            type: "object"
            properties:
              status:
                type: "boolean"
                example: false
              message:
                type: "string"
                example: "supplierCode parameter is missing"
  /api/view/singlesupplier/{supplierCode}:
    get:
      tags:
      - "Suppliers"
      summary: "Find single supplier by supplier code"
      description: ""
      operationId: "getsSupplierByCode"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "supplierCode"
        in: "path"
        description: "Code of supplier that needs to be fetched"
        required: true
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              status:
                type: "boolean"
                example: true
              message:
                type: "string"
                example: "Supplier found!!"
              data:
                $ref: "#/definitions/Supplier"
        "500":
          description: "Invalid input"
          schema:
            type: "object"
            properties:
              status:
                type: "boolean"
                example: false
              message:
                type: "string"
                example: "Supplier not found!!"
              data:
                type: "object"
        "400":
          description: "parameter is missing"
          schema:
            type: "object"
            properties:
              status:
                type: "boolean"
                example: false
              message:
                type: "string"
                example: "supplierCode parameter is missing"
...
definitions:
...
  Supplier:
    type: "object"
    properties:
      _partition: 
        type: "string"
        example: "101"
      userId:
        type: "string"
        #format: "objectId"
        example: "6048708c946251d02c88f329"
      contact: 
        type: "object"
        properties:
          address: 
            type:
            properties:
                addressLine1:
                  type: "string"
                  example: "PUNE"
                addressLine2:
                  type: "string"
                  example: "BANNER"
                city:
                  type: "string"
                  example: "PUNE"
                country:
                  type: "string"
                  example: "INDIA"
                isoCountry:
                  type: "string"
                  example: "phl"
                district:
                  type: "string"
                  example: "PUNE"
                postalCode:
                  type: "string"
                  example: "784512"
                directions:
                  type: "string"
                  example: "WEST"
                landmark:
                  type: "string"
                  example: "NEW GALAXY MALL"
                region:
                  type: "string"
                  example: "EAST"
                town:
                  type: "string"
                  example: "PUNE"
                zip:
                  type: "string"
                  example: "PUNE"            
          email:
            type: "string"
            example: "email@gmail.com"
          phone:
            type: "string"
            example: "9999999999"
      supplierCode:
        type: "string"
        required: true
        example: "ALKEM"
      deliveryFee:
        type: "integer"
        format: "decimal"
        example: 100.00
      isoCountry:
        type: "string"
        example: "phl"
      lastProductSeq:
        type: "string"
        example: "45A"
      supplierName:
        type: "string"
        example: "ALKEM"
      catalogTags:
        type: "array"
        xml:
          name: "tag"
          wrapped: true
        items:
          type: "string"
      usdPrice:
         type: "integer"
         format: "decimal"
    xml:
      name: "Supplier"
...